requirements.txt
gitpython
pyyaml
pluginbase
docker
python-slugify
transitions